+++ apps/admin/src/components/breads/bread-create-dialog.tsx
@@ -6,6 +6,7 @@ import {
   DialogHeader,
   DialogTitle,
   DialogTrigger,
+  ScrollArea,
 } from '@appabbang/ui';
 
 import { useBreadsCreateMutation } from '@/hooks/use-breads';
@@ -23,13 +24,15 @@ export function BreadCreateDialog() {
         onInteractOutside={(e) => {
           e.preventDefault();
         }}
-        className="sm:max-w-xl overflow-y-auto max-h-11/12"
+        className="sm:max-w-xl h-fit p-0"
       >
-        <DialogHeader>
-          <DialogTitle>메뉴등록</DialogTitle>
-        </DialogHeader>
-        <DialogDescription>메뉴를 등록해주세요</DialogDescription>
-        <BreadForm submitFn={breadsCreateMutation} />
+        <ScrollArea className="h-[700px] p-6">
+          <DialogHeader>
+            <DialogTitle>메뉴등록</DialogTitle>
+          </DialogHeader>
+          <DialogDescription>메뉴를 등록해주세요</DialogDescription>
+          <BreadForm submitFn={breadsCreateMutation} />
+        </ScrollArea>
       </DialogContent>
     </Dialog>
   );
