+++ apps/admin/src/components/breads/bread-form.tsx
@@ -107,7 +107,7 @@ function BreadForm({ submitFn, currentValues, no }: BreadFormProps) {
 
   return (
     <Form {...form}>
-      <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8 ">
+      <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8 py-4">
         <FormField
           control={form.control}
           name="name"
